interger is too small changed to big int on phone number.

+++ data/migrations/20201014204709_create_bare_bones_tables.js
@@ -4,7 +4,7 @@ exports.up = async (knex) => {
     table.string('user_id').references('id').inTable('profiles');
     table.string('given_name').notNull();
     table.string('family_name').notNull();
-    table.integer('phone-number').notNull();
+    table.bigint('phone-number').notNull();
     table.string('address').notNull();
     table.string('city').notNull();
     table.string('state').notNull();
@@ -16,7 +16,7 @@ exports.up = async (knex) => {
     table.string('user_id').references('id').inTable('profiles');
     table.string('given_name').notNull();
     table.string('family_name').notNull();
-    table.integer('phone_number').notNull();
+    table.bigint('phone_number').notNull();
     table.string('address');
     table.string('city');
     table.string('state');
